Refactor muscle workflow and remove deprecated operators
- Updated `Muscle_Preview_Update_Op` to remove debug print statements and streamline curve change detection.
- Removed `Muscle_Validation_Op`, `Muscle_Cleanup_Op`, `Muscle_Debug_Alignment_Op`, and other related operators to simplify the codebase.
- Consolidated muscle creation logic into `core_operators.py`, enhancing maintainability and organization.
- Removed `muscle_creation.py`, `muscle_modifications.py`, `muscle_parameters.py`, and `muscle_selection.py` as part of the cleanup process.
- Updated the UI panel in `myoGenerator_panel.py` to reflect the removal of validation and utility options.
- Added core operators for muscle name submission, selection, and parameter calculation to support the new workflow.
