--- AddonFolder/muscle_selection.py
@@ -1,92 +1,0 @@
-import bpy
-from .muscle_utilities import select_and_edit_object, create_mesh_from_selected_faces
-
-class Muscle_Name_Submition(bpy.types.Operator):
-    bl_idname = "view3d.submit_button"
-    bl_label = "Submit Muscle Name"
-    bl_description = "Submit Muscle Name"
-
-    def execute(self, context):
-        objName = bpy.context.scene.muscle_Name
-
-        # Check if the "muscles" collection exists
-        if "muscles" not in bpy.data.collections:
-            # Create the "muscles" collection
-            muscles_collection = bpy.data.collections.new("muscles")
-            bpy.context.scene.collection.children.link(muscles_collection)
-        else:
-            muscles_collection = bpy.data.collections["muscles"]
-
-        # Check if a collection with the name objName already exists inside "muscles"
-        if objName in muscles_collection.children:
-            self.report({'WARNING'}, f"Muscle '{objName}' already exists, choose a different name.")            
-            return {'CANCELLED'}
-        else:
-            # Create a new collection inside "muscles" with the name objName
-            new_collection = bpy.data.collections.new(objName)
-            muscles_collection.children.link(new_collection)
-
-        return {'FINISHED'}
-
-class Select_Origin_Op(bpy.types.Operator):
-    bl_idname = "view3d.select_origin"
-    bl_label = "Select Origin"
-    bl_description = "Select Origin of the muscle"
-
-    def execute(self, context):
-        # Get the selected object in origin object
-        origin_object = bpy.context.scene.origin_object
-        select_and_edit_object(origin_object)
-        return {'FINISHED'}
-
-class Select_Insertion_Op(bpy.types.Operator):
-    bl_idname = "view3d.select_insertion"
-    bl_label = "Select Insertion"
-    bl_description = "Select Insertion of the muscle"
-
-    def execute(self, context):
-        # Get the selected object in insertion object
-        insertion_object = bpy.context.scene.insertion_object
-        select_and_edit_object(insertion_object)
-        return {'FINISHED'}
-
-class Submit_Origin_Op(bpy.types.Operator):
-    bl_idname = "view3d.submit_origin"
-    bl_label = "Submit Origin"
-    bl_description = "Submit Origin of the muscle"
-
-    @classmethod
-    def poll(cls, context):
-        return context.object is not None and context.object.mode == 'EDIT'
-
-
-    def execute(self, context):
-        create_mesh_from_selected_faces(self,"origin")
-        return {'FINISHED'}
-
-class Submit_Insertion_Op(bpy.types.Operator):
-    bl_idname = "view3d.submit_insertion"
-    bl_label = "Submit Insertion"
-    bl_description = "Submit Insertion of the muscle"
-
-    @classmethod
-    def poll(cls, context):
-        return context.object is not None and context.object.mode == 'EDIT'
-
-    def execute(self, context):
-        create_mesh_from_selected_faces(self,"insertion")
-        return {'FINISHED'}
-
-class Next_Muscle_Op(bpy.types.Operator):
-    bl_idname = "view3d.next_muscle"
-    bl_label = "Next Muscle"
-    bl_description = "Start the creation of the next muscle"
-
-    def execute(self, context):
-        # Clean the muscle name from the muscle panel
-        bpy.context.scene.muscle_Name = "Insert muscle name"
-        # Change to object mode
-        bpy.ops.object.mode_set(mode='OBJECT')
-        # Deselect all objects
-        bpy.ops.object.select_all(action='DESELECT')
-        return {'FINISHED'}
